refactor(wallet): .interface.ts => .repository.interface.ts
