+++ src/wallet/domain/interfaces/point-transaction.repository.interface.ts
