+++ src/wallet/domain/interfaces/user-balance.repository.interface.ts
