+++ src/wallet/domain/use-cases/charge-points.use-case.spec.ts
@@ -4,8 +4,8 @@ import {
   InvalidChargeAmountError,
   UserBalanceNotFoundError,
 } from "../exceptions/point.exceptions";
-import { UserBalanceRepositoryInterface } from "../interfaces/user-balance.repository";
-import { PointTransactionRepositoryInterface } from "../interfaces/point-transaction.repository";
+import { UserBalanceRepositoryInterface } from "../interfaces/user-balance.repository.interface";
+import { PointTransactionRepositoryInterface } from "../interfaces/point-transaction.repository.interface";
 import { v4 as uuidv4 } from "uuid";
 
 describe("ChargePointsUseCase", () => {
