+++ src/wallet/domain/use-cases/recover-points.use-case.spec.ts
@@ -6,8 +6,8 @@ import {
   PointTransactionNotFoundError,
   UserBalanceNotFoundError,
 } from "../exceptions/point.exceptions";
-import { UserBalanceRepositoryInterface } from "../interfaces/user-balance.repository";
-import { PointTransactionRepositoryInterface } from "../interfaces/point-transaction.repository";
+import { UserBalanceRepositoryInterface } from "../interfaces/user-balance.repository.interface";
+import { PointTransactionRepositoryInterface } from "../interfaces/point-transaction.repository.interface";
 import { v4 as uuidv4 } from "uuid";
 
 describe("RecoverPointsUseCase", () => {
