+++ src/wallet/domain/use-cases/use-points.use-case.spec.ts
@@ -5,8 +5,8 @@ import {
   InsufficientPointBalanceError,
   UserBalanceNotFoundError,
 } from "../exceptions/point.exceptions";
-import { UserBalanceRepositoryInterface } from "../interfaces/user-balance.repository";
-import { PointTransactionRepositoryInterface } from "../interfaces/point-transaction.repository";
+import { UserBalanceRepositoryInterface } from "../interfaces/user-balance.repository.interface";
+import { PointTransactionRepositoryInterface } from "../interfaces/point-transaction.repository.interface";
 import { v4 as uuidv4 } from "uuid";
 
 describe("UsePointsUseCase", () => {
