+++ src/wallet/domain/use-cases/use-points.use-case.ts
@@ -5,8 +5,8 @@ import {
   InsufficientPointBalanceError,
   UserBalanceNotFoundError,
 } from "../exceptions/point.exceptions";
-import { PointTransactionRepositoryInterface } from "../interfaces/point-transaction.repository";
-import { UserBalanceRepositoryInterface } from "../interfaces/user-balance.repository";
+import { PointTransactionRepositoryInterface } from "../interfaces/point-transaction.repository.interface";
+import { UserBalanceRepositoryInterface } from "../interfaces/user-balance.repository.interface";
 
 export interface UsePointsUseCaseCommand {
   userId: string;
